Add -i parameter to provide a workaround for https://github.com/locustio/locust/issues/1085
Also move the checks parameters into __init__.py
